Optimize delta compression and improve concurrency safety
This release includes memory optimization for full frame compression and
multiple bug fixes for production stability.

Memory Optimization:
- Add compressed buffer reuse in delta encoder to eliminate allocations
- Reduce GC pressure during full frame streaming (0 allocs/op achieved)
- Implement proper buffer cleanup in ReleaseMemory()

Concurrency Fixes:
- Add mutex protection for JWT secret access (fixes race condition)
- Fix EventScanner goroutine cleanup with context cancellation
- Make Close() methods idempotent for FramebufferReader and EventScanner
- Fix rate parameter handling to be per-client instead of global

Error Handling Improvements:
- Add proper error handling for xochitl process discovery
- Clear framebuffer on read errors to prevent stale data
- Add nil checks for sync.Pool type assertions
- Fix timeout handling in pubsub to avoid ticker leak

Testing:
- Add race condition tests for JWT manager
- Add concurrent client tests for stream handler
- Add buffer reuse validation tests
- Add pool handling safety tests

Dependencies:
- Update GitHub Actions to use checkout@v4 and setup-go@v5
- Update Go version to 1.26

+++ internal/eventhttphandler/gesture_handler.go
@@ -33,7 +33,7 @@ type GestureHandler struct {
 }
 
 type gesture struct {
-	leftDistance, rightDistance, upDistance, downDistance int
+	leftDistance, rightDistance, upDistance, downDistance int64
 }
 
 func (g *gesture) MarshalJSON() ([]byte, error) {
@@ -44,7 +44,7 @@ func (g *gesture) String() string {
 	return fmt.Sprintf("Left: %v, Right: %v, Up: %v, Down: %v", g.leftDistance, g.rightDistance, g.upDistance, g.downDistance)
 }
 
-func (g *gesture) sum() int {
+func (g *gesture) sum() int64 {
 	return g.leftDistance + g.rightDistance + g.upDistance + g.downDistance
 }
 
@@ -113,9 +113,9 @@ func (h *GestureHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 				}
 				distance := event.Value - lastEventX.Value
 				if distance < 0 {
-					currentGesture.rightDistance += -int(distance)
+					currentGesture.rightDistance += -int64(distance)
 				} else {
-					currentGesture.leftDistance += int(distance)
+					currentGesture.leftDistance += int64(distance)
 				}
 				lastEventX = event
 			case codeYAxis:
@@ -126,9 +126,9 @@ func (h *GestureHandler) ServeHTTP(w http.ResponseWriter, r *http.Request) {
 				}
 				distance := event.Value - lastEventY.Value
 				if distance < 0 {
-					currentGesture.upDistance += -int(distance)
+					currentGesture.upDistance += -int64(distance)
 				} else {
-					currentGesture.downDistance += int(distance)
+					currentGesture.downDistance += int64(distance)
 				}
 				lastEventY = event
 			}
